backend/requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Mail==0.10.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.1.2
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-CORS==3.0.10
Celery==5.4.0
redis==5.0.1
